Remove ohlc_window_seconds in trade_producer
The ohlc_window_seconds parameter in trade_producer is not used, so
we remove it.
We also adjusted the trade_producer Dockerfile to copy only
the pyproject.toml, poetry.lock and src/*.py files.

This fix should remove bug when you have an `.venv` directory with
your virtualenv, as it messes things up when you `poetry install`
the dependencies inside the container.