End of live coding session 2024-06-10
This commit adds a `buffer_size parameter to the kafka_to_feature_store.
This parameter controls the number of messages that are saved in each
.insert() call to the feature group.

We save in batches instead of one-by-one to avoid the overhead of
inserting each message individually.

Warning! The code has at least 2 problems

Problem 1) Feature are saved to the online store, not to the offline
store. This needs to be fixed.

Problem 2) We don't handle the case where the last batch of data is
half-full. At the moment, we simply do not save it to the feature store,
so we are losing it. We need to fix this somehow.