Fix bug trade_producer in historical mode
The trade_producer was not working in historical mode because it was
fetching the same batch of trades over and over again. This was due to
the fact that we were not updating the last trade id after fetching the
trades. This commit fixes this issue by updating the `since` parameter
in the GET request.