Fetch data for many pairs at one from Kraken REST API
Added a new class called `KrakenRESTAPIMultipleProducts` with the same
interface as our previous `KrakenRESTAPI` class, which means it has 2
methods

- `get_trades()`
- `is_done()`

but this time we can fetch data for many pairs at once.

Under the hood, this new class is using a list of `KrakenRESTAPI` objects
to fetch data for each pair, and combining the results into a single
trade list output.