Overwrite the timestamp parameter in trade_producer
Quix Streams window operations use the kafka timestamp to do the
aggregations. This is fine when we work with live data, but it is not
ok when we want to transform historical data.

To fix this we decided to overwrite the timestamp parameter in the
trade message. We do this in the trade_producer.

We also extracted the consumer_group from the `trade_to_ohlc` service
as a config parameter.