Fix `trade_to_ohlc` to emit OHLC at the end of window
Before this commit, the `trade_to_ohlc` script would emit an ohlc candle
after processing each trade. This way, we would have a candle for each
trade, which is not what we want. We are not interested in all these
intermediate candles, but only in the candles at the end of the window.

Thanks Anton Javelosa for finding this bug!