Add Trade type to websocket.py and rest.py classes
With this Pydantic type we remove inconsistencies in the way
we created trade dictionaries in the websocket and rest classes.

With Pydantic objects we get type validation out of the box, so
we catch these errors right away.

We also created 2 .env files (live.env, historical.env) to encapsulate
all the configuration parameters we need to run the trade_producer
service either
> with live data coming the websocket, or
> with historical data coming from the rest api