Create new consumer group in kafka_to_feature_store
We added a new configuration parameter called `create_new_consumer_group`
in the `kafka_to_feature_store` to allow users to create a new consumer
so this way we make sure we re-process all the messages in the `ohlc_historical`
topic.

This is something we would like to do when for some reason one of the
jobs writing features to Hopsworks fails.

By passing the `create_new_consumer_group=True` we force the re-processing
of all messages.